Generate instruction templates as C structs
Any instruction-specific info is now generated as constant C structures,
rather than different executable code for each instruction. To achieve
this, a lot of logic has been moved from the generator to the C code.
The result is a drastic improvent wrt. compile times and code size,
in both source and binary form.